requirement.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
numpy==1.19.1
pandas==1.1.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
